pyproject.toml
[tool.poetry]
name = "tiktok-api-integration"
version = "1.0.0"
description = "Social Media Market Analysis Tool - TikTok API Integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
requests = "^2.31.0"
httpx = "^0.25.1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.7"
pydantic = {extras = ["email"], version = "^2.11.7"}
pydantic-settings = "^2.0.3"
cryptography = "^41.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pandas = "^2.1.3"
numpy = "^1.26.2"
redis = "^5.0.1"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.10.1"
pylint = "^3.0.2"
mypy = "^1.7.0"
ipython = "^8.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | alembic
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["app"]
skip_glob = ["*/migrations/*", "*/alembic/*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
namespace_packages = true
exclude = [
    "alembic/",
    "tests/",
]

[tool.pylint]
max-line-length = 88
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "W0212",  # protected-access
    "W0622",  # redefined-builtin
]
ignore = ["alembic", "migrations"]
extension-pkg-whitelist = "pydantic"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/alembic/*",
    "*/__init__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"